GenomicRegionSearch - fixed npr
Former-commit-id: 86aead8c3b5e50d7565926339b3dff214989e691

+++ bio/webapp/src/org/intermine/bio/web/logic/GenomicRegionSearchQueryRunner.java
@@ -418,9 +418,9 @@ class ValueComparator implements Comparator
 
     @Override
     public int compare(Object a, Object b) {
-        if ((Integer) base.get(a) < (Integer) base.get(b)) {
+        if (base.get(a) < base.get(b)) {
             return 1;
-        } else if ((Integer) base.get(a) == (Integer) base.get(b)) {
+        } else if (base.get(a) == base.get(b)) {
             return -1;
         } else {
             return -1;
